pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voicetype"
version = "0.1.0"
description = "Real-time voice-to-text input tool with AI-powered text cleanup"
readme = "README.md"
license = {text = "AGPL-3.0"}
requires-python = ">=3.10"
authors = [
    {name = "VoiceType Contributors"},
]
keywords = ["voice", "typing", "asr", "speech-to-text", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.10.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "sherpa-onnx",  # Lightweight speaker recognition (~15MB)
    "websockets>=13.0,<15.0",
    "pynput>=1.7.6",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
standalone = ["pystray>=0.19.0", "Pillow>=10.0.0"]
win = ["pywin32>=306"]
mac = ["pyobjc-framework-Cocoa>=10.0"]

[project.scripts]
voicetype = "voicetype.main:run"
download-sherpa-models = "scripts.download_sherpa_models:download_models"

[project.urls]
Homepage = "https://github.com/tmwgsicp/voicetype"
Repository = "https://github.com/tmwgsicp/voicetype"

[tool.setuptools.packages.find]
include = ["voicetype*"]
